feat: new about page + layout
