+++ frontend/scripts/react-components/about/about.component.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+// import PropTypes from 'prop-types';
+// import Link from 'redux-first-router-link';
+import Hero from 'react-components/shared/hero.component';
+
+function About() {
+  return (
+    <div className="c-about">
+      <Hero className="-read-only" />
+    </div>
+  );
+}
+
+export default About;
